requirements.txt
# FastAPI y Servidor ASGI
fastapi==0.111.0
uvicorn==0.29.0

# Socket.IO y su motor (versiones compatibles)
python-socketio==5.11.2
python-engineio==4.9.1

# LangChain - Versiones estables y compatibles
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langgraph==0.2.10

# Variables de entorno
python-dotenv==1.0.1

# Para manejar CORS y multipart
python-multipart==0.0.9

# Logging mejorado
colorlog==6.8.2